requirements.txt
discord.py==2.3.2
cohere==5.5.8
flask==3.0.0